+++ tests/validate.js
@@ -1,6 +1,6 @@
 const { resolve } = require("path");
-const createGeneratedApiInfos = require("./utils/createGeneratedApiInfos");
-const validateGeneratedModule = require("./utils/validateGeneratedModule");
+const createGeneratedApiInfos = require("./helpers/createGeneratedApiInfos");
+const validateGeneratedModule = require("./helpers/validateGeneratedModule");
 
 const v2ApiPaths = createGeneratedApiInfos(resolve(__dirname, "./generated/v2.0"));
 const v3ApiPaths = createGeneratedApiInfos(resolve(__dirname, "./generated/v3.0"));
@@ -10,8 +10,8 @@ let hasErrors = false;
 [
   ...v2ApiPaths,
   ...v3ApiPaths,
-].forEach(path => {
-  const diagnostics = validateGeneratedModule(path)
+].forEach(pathToFile => {
+  const diagnostics = validateGeneratedModule({ pathToFile })
 
   if (!hasErrors) {
     hasErrors = !!diagnostics.length